making a simple game

+++ assets/shader/sprite.frag
@@ -1,7 +1,7 @@
 #version 460 core
 
 in vec2 v_uv;
-flat in float v_tex_index;
+flat in int v_tex_index;
 
 out vec4 FragColor;
 
@@ -9,7 +9,7 @@ out vec4 FragColor;
 uniform sampler2D u_textures[16];
 
 void main() {
-    int index = int(v_tex_index);
+    int index = v_tex_index;
 
     if (index < 0 || index >= 16) {
          FragColor = vec4(1.0, 0.0, 1.0, 1.0); // magenta = error
